phil: imporved, updated, refactored and commented

+++ Curriculum/08_philosophers/01_code/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: akekesi <akekesi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/11 19:06:49 by akekesi           #+#    #+#             */
-/*   Updated: 2023/09/09 05:46:59 by akekesi          ###   ########.fr       */
+/*   Updated: 2023/09/12 19:34:55 by akekesi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "header.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	n;
+	int	n;
 
 	n = 0;
 	while (str[n])
